# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Lifelog API
  description: 일정관리 기반 라이프로그 서비스 API
  version: 1.0.0

servers:
  - url: http://localhost:8080/api/v1
    description: 로컬 개발

tags:
  - name: auth
    description: 인증 (회원가입 / 로그인 / 토큰 재발급 / 로그아웃)
  - name: events
    description: 일정 관리 CRUD
  - name: expense-categories
    description: 가계부 카테고리 관리
  - name: expenses
    description: 가계부 내역 CRUD / 요약

# ------------------------------------------------------------------
# 보안: 모든 엔드포인트는 기본적으로 BearerAuth 필요
#        auth 엔드포인트만 security: [] 로 override
# ------------------------------------------------------------------
security:
  - BearerAuth: []

paths:

  # ================================================================
  # AUTH
  # ================================================================
  /auth/signup:
    post:
      tags: [auth]
      summary: 회원가입
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: 가입 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_UserResponse'
        '409':
          description: 이메일 중복
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'

  /auth/login:
    post:
      tags: [auth]
      summary: 로그인
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 로그인 성공 — access/refresh 토큰 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_TokenResponse'
        '401':
          description: 이메일 또는 비밀번호 불일치
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'

  /auth/refresh:
    post:
      tags: [auth]
      summary: 액세스 토큰 재발급
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: 새 액세스 토큰 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_TokenResponse'
        '401':
          description: refresh 토큰 만료 또는 무효
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'

  /auth/logout:
    post:
      tags: [auth]
      summary: 로그아웃 (refresh 토큰 무효화)
      responses:
        '200':
          description: 로그아웃 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Empty'

  # ================================================================
  # EVENTS
  # ================================================================
  /events:
    get:
      tags: [events]
      summary: 일정 목록 조회
      parameters:
        - name: from
          in: query
          description: 조회 시작일 (ISO 8601, 포함)
          schema:
            type: string
            format: date-time
            example: '2025-01-01T00:00:00'
        - name: to
          in: query
          description: 조회 종료일 (ISO 8601, 포함)
          schema:
            type: string
            format: date-time
            example: '2025-01-31T23:59:59'
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: 일정 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponse_EventSummary'

    post:
      tags: [events]
      summary: 일정 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequest'
      responses:
        '201':
          description: 생성된 일정
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_EventResponse'
        '400':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'

  /events/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64

    get:
      tags: [events]
      summary: 일정 단건 조회
      responses:
        '200':
          description: 일정 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_EventResponse'
        '404':
          description: 일정 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'

    put:
      tags: [events]
      summary: 일정 수정
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequest'
      responses:
        '200':
          description: 수정된 일정
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_EventResponse'
        '403':
          description: 본인 일정 아님
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'
        '404':
          description: 일정 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'

    delete:
      tags: [events]
      summary: 일정 삭제
      responses:
        '200':
          description: 삭제 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Empty'
        '403':
          description: 본인 일정 아님
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'
        '404':
          description: 일정 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'

  # ================================================================
  # EXPENSE CATEGORIES
  # ================================================================
  /expense-categories:
    get:
      tags: [expense-categories]
      summary: 가계부 카테고리 목록 조회
      description: 페이지네이션 없음 (사용자당 최대 100개). type → name 순 정렬.
      parameters:
        - name: type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ExpenseType'
      responses:
        '200':
          description: 카테고리 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ExpenseCategoryList'
        '400':
          description: 잘못된 파라미터
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'

    post:
      tags: [expense-categories]
      summary: 가계부 카테고리 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseCategoryCreateRequest'
      responses:
        '201':
          description: 생성된 카테고리
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ExpenseCategoryResponse'
        '400':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'
        '409':
          description: 같은 유형 내 이름 중복 또는 사용자당 카테고리 100개 초과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'

  /expense-categories/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64

    put:
      tags: [expense-categories]
      summary: 가계부 카테고리 이름 변경
      description: 유형(type)은 변경할 수 없다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseCategoryUpdateRequest'
      responses:
        '200':
          description: 수정된 카테고리
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ExpenseCategoryResponse'
        '400':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'
        '403':
          description: 본인 카테고리 아님
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'
        '404':
          description: 카테고리 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'
        '409':
          description: 같은 유형 내 이름 중복
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'

    delete:
      tags: [expense-categories]
      summary: 가계부 카테고리 삭제
      responses:
        '200':
          description: 삭제 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Empty'
        '403':
          description: 본인 카테고리 아님
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'
        '404':
          description: 카테고리 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'
        '409':
          description: 해당 카테고리를 사용하는 내역이 있어 삭제 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'

  # ================================================================
  # EXPENSES
  # ================================================================
  /expenses:
    get:
      tags: [expenses]
      summary: 가계부 내역 목록 조회
      description: 거래일 내림차순, 같은 날짜는 id 내림차순.
      parameters:
        - name: from
          in: query
          description: 조회 시작일 (포함)
          schema:
            type: string
            format: date
            example: '2026-09-01'
        - name: to
          in: query
          description: 조회 종료일 (포함). from보다 이르면 400
          schema:
            type: string
            format: date
            example: '2026-09-30'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/ExpenseType'
        - name: categoryId
          in: query
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: 내역 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponse_ExpenseSummary'
        '400':
          description: 잘못된 파라미터 (from > to, size 범위 초과 등)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'

    post:
      tags: [expenses]
      summary: 가계부 내역 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseRequest'
      responses:
        '201':
          description: 생성된 내역
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ExpenseResponse'
        '400':
          description: 유효성 오류 또는 카테고리 유형과 내역 유형 불일치
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'
        '403':
          description: 본인 카테고리 아님
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'
        '404':
          description: 카테고리 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'

  /expenses/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64

    get:
      tags: [expenses]
      summary: 가계부 내역 단건 조회
      responses:
        '200':
          description: 내역 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ExpenseResponse'
        '403':
          description: 본인 내역 아님
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'
        '404':
          description: 내역 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'

    put:
      tags: [expenses]
      summary: 가계부 내역 수정
      description: 카테고리 변경 허용. 카테고리 유형과 내역 유형이 일치해야 한다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseRequest'
      responses:
        '200':
          description: 수정된 내역
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ExpenseResponse'
        '400':
          description: 유효성 오류 또는 카테고리 유형과 내역 유형 불일치
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'
        '403':
          description: 본인 내역/카테고리 아님
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'
        '404':
          description: 내역 또는 카테고리 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'

    delete:
      tags: [expenses]
      summary: 가계부 내역 삭제
      responses:
        '200':
          description: 삭제 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Empty'
        '403':
          description: 본인 내역 아님
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'
        '404':
          description: 내역 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'

  /expenses/summary/monthly:
    get:
      tags: [expenses]
      summary: 월별 수입/지출 요약
      description: 기간 내 모든 달을 포함하며 내역이 없는 달은 0으로 채운다. 최대 12개월.
      parameters:
        - name: from
          in: query
          required: true
          description: 시작 월 (yyyy-MM)
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
            example: '2026-01'
        - name: to
          in: query
          required: true
          description: 종료 월 (yyyy-MM, 포함)
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
            example: '2026-09'
      responses:
        '200':
          description: 월별 요약
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_MonthlySummaryResponse'
        '400':
          description: 파라미터 누락/형식 오류, from > to, 12개월 초과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'

  /expenses/summary/by-category:
    get:
      tags: [expenses]
      summary: 카테고리별 합계
      description: 금액 내림차순. 최대 366일.
      parameters:
        - name: type
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ExpenseType'
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
            example: '2026-09-01'
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
            example: '2026-09-30'
      responses:
        '200':
          description: 카테고리별 합계
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_CategorySummaryResponse'
        '400':
          description: 파라미터 누락/형식 오류, from > to, 366일 초과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'

# ------------------------------------------------------------------
# COMPONENTS
# ------------------------------------------------------------------
components:

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:

    # ── 공통 응답 envelope ──────────────────────────────────────────
    ApiResponse_UserResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/UserResponse'
        error: { type: string, nullable: true, example: null }

    ApiResponse_TokenResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/TokenResponse'
        error: { type: string, nullable: true, example: null }

    ApiResponse_EventResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/EventResponse'
        error: { type: string, nullable: true, example: null }

    ApiResponse_Empty:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: { type: object, nullable: true, example: null }
        error: { type: string, nullable: true, example: null }

    ApiResponse_Error:
      type: object
      properties:
        success: { type: boolean, example: false }
        data: { type: object, nullable: true, example: null }
        error: { type: string, example: '이미 사용중인 이메일입니다.' }

    PagedResponse_EventSummary:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: array
          items:
            $ref: '#/components/schemas/EventSummary'
        error: { type: string, nullable: true, example: null }
        meta:
          $ref: '#/components/schemas/PageMeta'

    ApiResponse_ExpenseCategoryResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/ExpenseCategoryResponse'
        error: { type: string, nullable: true, example: null }

    ApiResponse_ExpenseCategoryList:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: array
          items:
            $ref: '#/components/schemas/ExpenseCategoryResponse'
        error: { type: string, nullable: true, example: null }

    ApiResponse_ExpenseResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/ExpenseResponse'
        error: { type: string, nullable: true, example: null }

    ApiResponse_MonthlySummaryResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/MonthlySummaryResponse'
        error: { type: string, nullable: true, example: null }

    ApiResponse_CategorySummaryResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/CategorySummaryResponse'
        error: { type: string, nullable: true, example: null }

    PagedResponse_ExpenseSummary:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: array
          items:
            $ref: '#/components/schemas/ExpenseSummary'
        error: { type: string, nullable: true, example: null }
        meta:
          $ref: '#/components/schemas/PageMeta'

    PageMeta:
      type: object
      properties:
        page: { type: integer, example: 0 }
        size: { type: integer, example: 20 }
        total: { type: integer, format: int64, example: 100 }
        totalPages: { type: integer, example: 5 }

    # ── Auth ────────────────────────────────────────────────────────
    SignupRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 8
          example: 'P@ssw0rd!'
        name:
          type: string
          example: 홍길동

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string

    UserResponse:
      type: object
      properties:
        id: { type: integer, format: int64, example: 1 }
        email: { type: string, example: user@example.com }
        name: { type: string, example: 홍길동 }
        createdAt: { type: string, format: date-time }

    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT — 유효기간 15분
        refreshToken:
          type: string
          description: JWT — 유효기간 7일
        tokenType:
          type: string
          example: Bearer

    # ── Events ──────────────────────────────────────────────────────
    EventRequest:
      type: object
      required: [title, startAt]
      properties:
        title:
          type: string
          maxLength: 200
          example: 팀 미팅
        description:
          type: string
          nullable: true
          example: 주간 스프린트 리뷰
        startAt:
          type: string
          format: date-time
          example: '2025-02-01T10:00:00'
        endAt:
          type: string
          format: date-time
          nullable: true
          example: '2025-02-01T11:00:00'
        allDay:
          type: boolean
          default: false
        location:
          type: string
          nullable: true
          example: 강남구 선릉
        color:
          type: string
          nullable: true
          description: HEX 색상 코드
          example: '#4A90E2'
        eventCategory:
          type: string
          enum: [PERSONAL, WORK, REMINDER, OTHER]
          default: PERSONAL

    EventResponse:
      type: object
      properties:
        id: { type: integer, format: int64, example: 1 }
        title: { type: string, example: 팀 미팅 }
        description: { type: string, nullable: true }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time, nullable: true }
        allDay: { type: boolean }
        location: { type: string, nullable: true }
        color: { type: string, nullable: true }
        eventCategory:
          type: string
          enum: [PERSONAL, WORK, REMINDER, OTHER]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    EventSummary:
      description: 목록 조회용 — 상세 필드 제외
      type: object
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time, nullable: true }
        allDay: { type: boolean }
        color: { type: string, nullable: true }
        eventCategory:
          type: string
          enum: [PERSONAL, WORK, REMINDER, OTHER]

    # ── Expenses ────────────────────────────────────────────────────
    ExpenseType:
      type: string
      enum: [EXPENSE, INCOME]

    ExpenseCategoryCreateRequest:
      type: object
      required: [type, name]
      properties:
        type:
          $ref: '#/components/schemas/ExpenseType'
        name:
          type: string
          maxLength: 50
          description: 앞뒤 공백은 제거되어 저장됨. (사용자, 유형, 이름) 유니크
          example: 식비

    ExpenseCategoryUpdateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 50
          example: 외식

    ExpenseCategoryResponse:
      type: object
      properties:
        id: { type: integer, format: int64, example: 1 }
        type:
          $ref: '#/components/schemas/ExpenseType'
        name: { type: string, example: 식비 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    ExpenseRequest:
      type: object
      required: [type, categoryId, amount, transactionDate]
      properties:
        type:
          $ref: '#/components/schemas/ExpenseType'
        categoryId:
          type: integer
          format: int64
          description: 카테고리 유형이 type과 일치해야 함
          example: 1
        amount:
          type: integer
          format: int64
          minimum: 1
          maximum: 99999999999
          description: 원(KRW) 단위 정수
          example: 12000
        transactionDate:
          type: string
          format: date
          example: '2026-09-28'
        description:
          type: string
          maxLength: 200
          nullable: true
          example: 점심
        memo:
          type: string
          nullable: true

    ExpenseResponse:
      type: object
      properties:
        id: { type: integer, format: int64, example: 1 }
        type:
          $ref: '#/components/schemas/ExpenseType'
        categoryId: { type: integer, format: int64, example: 1 }
        categoryName: { type: string, example: 식비 }
        amount: { type: integer, format: int64, example: 12000 }
        transactionDate: { type: string, format: date }
        description: { type: string, nullable: true }
        memo: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    ExpenseSummary:
      description: 목록 조회용 — memo 및 시각 필드 제외
      type: object
      properties:
        id: { type: integer, format: int64 }
        type:
          $ref: '#/components/schemas/ExpenseType'
        categoryId: { type: integer, format: int64 }
        categoryName: { type: string }
        amount: { type: integer, format: int64 }
        transactionDate: { type: string, format: date }
        description: { type: string, nullable: true }

    MonthlySummaryResponse:
      type: object
      properties:
        from: { type: string, example: '2026-01', description: yyyy-MM }
        to: { type: string, example: '2026-09', description: yyyy-MM }
        totalIncome: { type: integer, format: int64 }
        totalExpense: { type: integer, format: int64 }
        net: { type: integer, format: int64, description: totalIncome - totalExpense }
        months:
          type: array
          items:
            $ref: '#/components/schemas/MonthlyItem'

    MonthlyItem:
      type: object
      properties:
        yearMonth: { type: string, example: '2026-09', description: yyyy-MM }
        income: { type: integer, format: int64 }
        expense: { type: integer, format: int64 }
        net: { type: integer, format: int64 }

    CategorySummaryResponse:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/ExpenseType'
        from: { type: string, format: date }
        to: { type: string, format: date }
        total: { type: integer, format: int64 }
        categories:
          type: array
          description: amount 내림차순
          items:
            $ref: '#/components/schemas/CategoryItem'

    CategoryItem:
      type: object
      properties:
        categoryId: { type: integer, format: int64 }
        categoryName: { type: string }
        amount: { type: integer, format: int64 }
        count: { type: integer, format: int64 }